requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
exceptiongroup==1.2.2
fast-depends==2.4.12
fastapi==0.115.11
faststream==0.5.35
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.1.0
pamqp==3.3.0
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
